backend/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langgraph-checkpoint-sqlite
openai
langsmith
httpx
spacy
requests
Pillow
python-dotenv
convex
qdrant-client
